requirements.txt
# Strands Agents Ecosystem Requirements
# ====================================

# Core framework
strands-agents>=0.1.7
strands-tools>=0.1.5

# Environment and configuration
python-dotenv>=1.0.0

# AWS integration
boto3>=1.34.0
botocore>=1.34.0